fix(wrapped): the poster carries the right name, the reel ends on your terms
The share poster — "the only thing in the reel a stranger will ever
see" — was signed "SimpMusic", and the saved file was
`SimpMusic_Wrapped_…`. Both say Orinify. Card 01 and THE REGULAR use
`daysElapsed` with "so far" copy while the year is unfinished.

The reel closed itself 30 seconds after the last card. It now stops
there and offers Watch again and Done, so the story ends when the
listener says so; the share pill is 48dp. The `HttpClient` created for
the reel's theme is closed on dispose — it leaked an engine per visit.

TalkBack: the close and back buttons are named, the pager announces
"Card N of M" with next/previous actions instead of invisible
half-screen taps, and the ring, radar and comparison bars speak the
numbers the cards already print. Locale: the eyebrows uppercase with
the current locale (the Turkish-i hazard one card already documented),
AM/PM and the decade suffix come from resources. The poster's rank rows
grow with the font instead of clipping; the top-tracks list scrolls
when the card is short; the per-card gap constants live in
`WrappedTokens`. `formatCount` moves to the analytics package so both
screens group thousands the same way.

+++ composeApp/src/commonMain/kotlin/com/maxrave/simpmusic/ui/component/WrappedEntryCard.kt
@@ -28,7 +28,7 @@ import androidx.compose.ui.unit.dp
 import coil3.compose.AsyncImage
 import com.maxrave.simpmusic.ui.icon.PlayArrow
 import com.maxrave.simpmusic.ui.icon.SimpIcons
-import com.maxrave.simpmusic.ui.screen.home.wrapped.formatCount
+import com.maxrave.simpmusic.ui.screen.home.analytics.formatCount
 import com.maxrave.simpmusic.ui.screen.home.wrapped.wholeMinutes
 import com.maxrave.simpmusic.ui.theme.typo
 import com.maxrave.simpmusic.viewModel.WrappedYear
@@ -98,7 +98,7 @@ fun WrappedEntryCard(
                         stringResource(
                             Res.string.wrapped_entry_subtitle,
                             formatCount(wholeMinutes(wrapped.stats.listenedSeconds)),
-                            formatCount(wrapped.stats.distinctArtists),
+                            formatCount(wrapped.stats.distinctArtists.toLong()),
                         ),
                     style = typo().bodySmall,
                     color = MaterialTheme.colorScheme.onSurfaceVariant,
